Fix naming code
