+++ Assets/Scripts/GameManager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 23cf1a1c8cbe99f47bb2f2b6abedd02f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
