+++ Assets/Scripts/Trash/TrashController.cs
@@ -1,6 +1,7 @@
 using Agate.MVC.Core;
 using EcoTeam.EcoToss.ObjectPooling;
 using EcoTeam.EcoToss.PubSub;
+using EcoTeam.EcoToss.GameManager;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -50,7 +51,7 @@ namespace EcoTeam.EcoToss.Trash
             {
                 PublishSubscribe.Instance.Publish<MessageTrashSpawn>(new MessageTrashSpawn());
 
-                if (GameManager.Instance.IsWindSpawn)
+                if (GameManagerController.Instance.IsWindSpawn)
                 {
                     PublishSubscribe.Instance.Publish<MessageSetRandomPropetiesWindArea>(new MessageSetRandomPropetiesWindArea());
                 }
